chore: fix trivial clippy warnings in active code
- Remove unnecessary u16 cast in contact/service.rs
- Simplify error pattern matching in group/broadcast.rs

Remaining 8 warnings are in unimplemented file transfer code (core/file/),
which can be addressed when implementing that feature.

Main code compiles cleanly with cargo check --lib

+++ src-tauri/src/core/contact/service.rs
@@ -140,7 +140,7 @@ impl ContactService {
                 uid: u.uid,
                 nickname: u.nickname,
                 feiq_ip: u.feiq_ip,
-                feiq_port: u.feiq_port as u16,
+                feiq_port: u.feiq_port,
                 feiq_machine_id: u.feiq_machine_id,
                 avatar: u.avatar,
                 status: u.status,
